add chimerseq package directories
