+++ src/main/java/org/com/chimerdbv31/chimerkb/mapper/ChimerKbMapper.java
@@ -5,7 +5,7 @@ import org.com.chimerdbv31.chimerkb.vo.ChimerKBVo;
 import org.com.chimerdbv31.common.vo.ParamVo;
 import org.springframework.stereotype.Repository;
 
-@Repository(value = "ChimerKbMapper")
+@Repository(value = "chimerKbMapper")
 public interface ChimerKbMapper {
     public int getChimerKBTotalNumber(ParamVo param);
     public List<ChimerKBVo> getChimerKBResult(ParamVo param);
