+++ src/main/java/org/com/chimerdbv31/chimerseq/services/ChimerSeqService.java
@@ -3,23 +3,23 @@ package org.com.chimerdbv31.chimerseq.services;
 import java.util.List;
 import javax.annotation.Resource;
 import org.com.chimerdbv31.chimerseq.mapper.ChimerSeqMapper;
-import org.com.chimerdbv31.chimerseq.vo.CimrSeqVo;
+import org.com.chimerdbv31.chimerseq.vo.ChimerSeqVo;
 import org.com.chimerdbv31.common.vo.ParamVo;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
-@Service(value = "ChimerSeqService")
+@Service(value = "chimerSeqService")
 public class ChimerSeqService {	
     @SuppressWarnings("unused")
     private static final Logger logger = LoggerFactory.getLogger(ChimerSeqService.class);
 
-    @Resource(name = "ChimerSeqMapper")
+    @Resource(name = "chimerSeqMapper")
     private ChimerSeqMapper chimerSeqMapper;
 
-    public List<CimrSeqVo> getChimerSeqResult(ParamVo param) {
-        List<CimrSeqVo> result = null;
+    public List<ChimerSeqVo> getChimerSeqResult(ParamVo param) {
+        List<ChimerSeqVo> result = null;
         try {
             result = this.chimerSeqMapper.getChimerSeqResult(param);
         } catch (Exception e) {
@@ -38,8 +38,8 @@ public class ChimerSeqService {
         return result;
     }
 
-    public List<CimrSeqVo> getResult(ParamVo param) {
-        List<CimrSeqVo> result = null;
+    public List<ChimerSeqVo> getResult(ParamVo param) {
+        List<ChimerSeqVo> result = null;
         try {
             result = this.chimerSeqMapper.getResult(param);
         } catch (Exception e) {
@@ -48,8 +48,8 @@ public class ChimerSeqService {
         return result;
     }
 
-    public List<CimrSeqVo> getResultTest(ParamVo param) {
-        List<CimrSeqVo> result = null;
+    public List<ChimerSeqVo> getResultTest(ParamVo param) {
+        List<ChimerSeqVo> result = null;
         try {
             result = this.chimerSeqMapper.getResultTest(param);
         } catch (Exception e) {
