+++ src/main/java/org/com/chimerdbv31/dao/MainDao.java
@@ -8,7 +8,7 @@ package org.com.chimerdbv31.dao;
 
 import org.com.chimerdbv31.iface.MainDaoIF;
 import org.com.chimerdbv31.chimerkb.vo.ChimerKBVo;
-import org.com.chimerdbv31.chimerseq.vo.CimrSeqVo;
+import org.com.chimerdbv31.chimerseq.vo.ChimerSeqVo;
 import org.com.chimerdbv31.common.vo.ParamVo;
 import java.util.List;
 import org.mybatis.spring.SqlSessionTemplate;
@@ -26,37 +26,37 @@ import org.mybatis.spring.SqlSessionTemplate;
 public class MainDao implements MainDaoIF{
     private SqlSessionTemplate sqlTemplate;
     
-    @Override
-    public void setnSqlSessionTemplate(SqlSessionTemplate sqlSessionTemplate) {
-        this.sqlTemplate = sqlSessionTemplate;
-    }
+	@Override
+	public void setnSqlSessionTemplate(SqlSessionTemplate sqlSessionTemplate) {
+		this.sqlTemplate = sqlSessionTemplate;
+	}
 
 
-    @Override
-    public List<CimrSeqVo> getResult(ParamVo param) {
-        List<CimrSeqVo> result = null;
-        try {
-            result = sqlTemplate.selectList("main_mper.q1_result_for_search",param);
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-        return result;
-    }
+	@Override
+	public List<ChimerSeqVo> getResult(ParamVo param) {
+		List<ChimerSeqVo> result = null;
+		try {
+			result = sqlTemplate.selectList("main_mper.q1_result_for_search",param);
+		} catch (Exception e) {
+			System.out.println(e.getMessage());
+		}
+		return result;
+	}
+
+	@Override
+	public List<ChimerSeqVo> getResultTest(ParamVo param) {
+		List<ChimerSeqVo> result = null;
+		try {
+			result = sqlTemplate.selectList("main_mper.q2_result_for_search",param);
+		} catch (Exception e) {
+			System.out.println(e.getMessage());
+		}
+		return result;
+	}
 
     @Override
-    public List<CimrSeqVo> getResultTest(ParamVo param) {
-        List<CimrSeqVo> result = null;
-        try {
-            result = sqlTemplate.selectList("main_mper.q2_result_for_search",param);
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-        return result;
-    }
-
-    @Override
-    public List<CimrSeqVo> getChimerSeqResult(ParamVo param) {
-        List<CimrSeqVo> result = null;
+    public List<ChimerSeqVo> getChimerSeqResult(ParamVo param) {
+        List<ChimerSeqVo> result = null;
         try {
             result = sqlTemplate.selectList("main_mper.q10_result_of_chimerseq",param);
         } catch (Exception e) {
