+++ src/main/java/org/com/chimerdbv31/iface/MainDaoIF.java
@@ -9,7 +9,7 @@ package org.com.chimerdbv31.iface;
 
 
 import org.com.chimerdbv31.chimerkb.vo.ChimerKBVo;
-import org.com.chimerdbv31.chimerseq.vo.CimrSeqVo;
+import org.com.chimerdbv31.chimerseq.vo.ChimerSeqVo;
 import org.com.chimerdbv31.common.vo.MainTableVo;
 import org.com.chimerdbv31.common.vo.ParamVo;
 import java.util.List;
@@ -22,9 +22,9 @@ import org.mybatis.spring.SqlSessionTemplate;
 public interface MainDaoIF {
     public void setnSqlSessionTemplate(SqlSessionTemplate sqlSessionTemplate);
     public int getChimerSeqTotalNumber(ParamVo param);
-    public List<CimrSeqVo> getChimerSeqResult(ParamVo param);
+    public List<ChimerSeqVo> getChimerSeqResult(ParamVo param);
     public int getChimerKBTotalNumber(ParamVo param);
     public List<ChimerKBVo> getChimerKBResult(ParamVo param);
-    public List<CimrSeqVo> getResult(ParamVo param);
-    public List<CimrSeqVo> getResultTest(ParamVo param);
+    public List<ChimerSeqVo> getResult(ParamVo param);
+    public List<ChimerSeqVo> getResultTest(ParamVo param);
 }
